rename story page folder
